Add Day 11
Definately my worse performance yet.
The code looks horrible and takes a while to run.

There is probably some optimisation where one of the numbers in the input was chosen because of its cyclic nature.
The code can only capture cyclics if they go below 10.
Could potentially experiment with catching other cycles in a future commit.